refactored schema utils
