+++ notifiers/providers/popcornnotify.py
@@ -1,6 +1,6 @@
 from ..core import Provider, Response
 from ..utils import requests
-from ..utils.json_schema import one_or_more, list_to_commas
+from ..utils.schema.helpers import one_or_more, list_to_commas
 
 
 class PopcornNotify(Provider):
