+++ notifiers/utils/schema/formats.py
@@ -4,9 +4,7 @@ from datetime import datetime
 
 import jsonschema
 
-from .helpers import valid_file
-
-# todo reformat this module into a package and add test
+from notifiers.utils.helpers import valid_file
 
 # Taken from https://gist.github.com/codehack/6350492822e52b7fa7fe
 ISO8601 = re.compile(
@@ -19,49 +17,7 @@ ISO8601 = re.compile(
     r'(?P<sec>[0-5][0-9]([,.]\d{1,10})?))?)?'
     r'(?:Z|([\-+](?:([01][0-9])|(?:2[0123]))(:?(?:[0-5][0-9]))?))?)?))$'
 )
-
 E164 = re.compile(r'^\+?[1-9]\d{1,14}$')
-
-
-def one_or_more(schema: dict, unique_items: bool = True, min: int = 1, max: int = None) -> dict:
-    """
-    Helper function to construct a schema that validates items matching
-    `schema` or an array containing items matching `schema`.
-
-    :param schema: The schema to use
-    :param unique_items: Flag if array items should be unique
-    :param min: Correlates to ``minLength`` attribute of JSON Schema array
-    :param max: Correlates to ``maxLength`` attribute of JSON Schema array
-    """
-    multi_schema = {
-        'type': 'array',
-        'items': schema,
-        'minItems': min,
-        'uniqueItems': unique_items
-    }
-    if max:
-        multi_schema['maxItems'] = max
-    return {
-        'oneOf': [
-            multi_schema,
-            schema
-        ]
-    }
-
-
-def list_to_commas(list_of_args) -> str:
-    """
-    Converts a list of items to a comma separated list. If ``list_of_args`` is
-    not a list, just return it back
-
-    :param list_of_args: List of items
-    :return: A string representing a comma separated list.
-    """
-    if isinstance(list_of_args, list):
-        return ",".join(list_of_args)
-    return list_of_args
-
-
 format_checker = jsonschema.FormatChecker()
 
 
@@ -73,7 +29,7 @@ def is_iso8601(instance: str):
 
 
 @format_checker.checks('rfc2822', raises=ValueError)
-def is_iso8601(instance: str):
+def is_rfc2822(instance: str):
     if not isinstance(instance, str):
         return True
     return email.utils.parsedate(instance) is not None
